refactor(ion-identity-client): minor changes to files (move things here and there)
